requirements.txt
Flask>=2.3
numpy>=1.23
Pillow>=9.0
scikit-image>=0.19
